requirements/development.txt
-r base.txt

django-debug-toolbar
django-extensions
werkzeug

git+https://github.com/pashinin/fabric@6783d2f7ad7a449c97cd1ee0df9a2dbc29422fce#egg=fabric
